HYBRID_QR_BARCODE_COMMUNICATION_SYSTEM/requirements.txt
Flask
qrcode
python-barcode
Pillow
opencv-python
pyzbar
numpy
pytest
